requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
seaborn
Pillow
tqdm